Fix overflow issue when confirming trade

+++ src/map/lua/lua_baseentity.cpp
@@ -4060,8 +4060,8 @@ void CLuaBaseEntity::confirmTrade()
             CItem* PItem = PChar->TradeContainer->getItem(slotID);
             if (PItem)
             {
-                uint8 confirmedItems = PChar->TradeContainer->getConfirmedStatus(slotID);
-                auto  quantity       = (int32)std::min<uint32>(PChar->TradeContainer->getQuantity(slotID), confirmedItems);
+                uint32 confirmedItems = PChar->TradeContainer->getConfirmedStatus(slotID);
+                auto   quantity       = (int32)std::min<uint32>(PChar->TradeContainer->getQuantity(slotID), confirmedItems);
 
                 PItem->setReserve(PItem->getReserve() - quantity);
                 if (confirmedItems > 0)
